AML-62 Finished GetFeedAsync service method and implemented custom exception handling

+++ AmourLink.Recommendation/Data/Entities/Preference.cs
@@ -8,7 +8,7 @@ public class Preference : Entity
     public int MinAge { get; set; }
     public Guid UserId { get; set; }
     public required string Gender { get; set; }
-    public decimal DistanceRange { get; set; }
+    public int DistanceRange { get; set; }
     
     public required User User { get; set; }
 }
